requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.1.0
fonttools==4.55.3
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.2.3
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1